[issue #25]警告レベルを最大にして警告が表示されないようにする
- プロジェクトの警告レベルを最大に変更し、警告をエラーとするように設定
- プロジェクトのInclude Path等を絶対パスから相対パスに変更
- 数値型のsigned/ unsignedが合うように型修正
- STL、DirectX等はプリコンパイル済ヘッダに記載し警告をdisableに設定
- ファイルのincludeをファイルからの相対パスでなくソリューションファイルからの相対パスに変更
- `FloatPoint`構造体を廃止し、`POINT`構造体のジェネリック版を作成
- 未使用の引数を削除
- インターフェースクラスのデストラクタをvirtualで定義
- インスタンス変数と同名の引数、ローカル変数は変数名変更
- gitignoreをVS2017で作成したものに変更

+++ BestSteal_Replica/Drawing/Drawer.cpp
@@ -3,7 +3,7 @@
 #include <d3dx9tex.h>
 
 #include "Drawer.h"
-#include "../Drawing/IDRawable.h"
+#include "Drawing/IDRawable.h"
 
 
 #define D3DFVF_CUSTOMVERTEX (D3DFVF_XYZRHW | D3DFVF_DIFFUSE | D3DFVF_TEX1)
@@ -83,7 +83,7 @@ void Drawer::Draw() {
 	// 描画データ生成
 	DataTable<DrawingContext> contextTable;
 	contextTable.resize(pDrawables.size());
-	for (int i = 0; i < (int)pDrawables.size(); ++i) {
+	for (UINT16 i = 0; i < pDrawables.size(); ++i) {
 		pDrawables[i]->CreateDrawingContexts(&contextTable[i]);
 	}
